requirements.txt
streamlit
numpy
pandas
st-annotated-text
treelib
plotly
plotly-express
geopandas